refactor: extract thread input orchestrator
