+++ apps/codex-runtime/src/domain/threads/thread-service.ts
@@ -1,12 +1,9 @@
-import crypto from "node:crypto";
-
 import { and, desc, eq } from "drizzle-orm";
 
 import type { RuntimeDatabase } from "../../db/database.js";
 import {
   type ApprovalRow,
   approvals,
-  messages,
   type SessionRow,
   sessionEvents,
   sessions,
@@ -27,6 +24,7 @@ import type {
 } from "../sessions/types.js";
 import type { WorkspaceFilesystem } from "../workspaces/workspace-filesystem.js";
 import type { WorkspaceRegistry } from "../workspaces/workspace-registry.js";
+import { ThreadInputOrchestrator } from "./thread-input-orchestrator.js";
 import type {
   LatestResolvedRequestSummary,
   NotificationEvent,
@@ -186,17 +184,9 @@ function deriveThreadTitle(content: string) {
   return compact.length <= 48 ? compact : `${compact.slice(0, 45)}...`;
 }
 
-function generateMessageId() {
-  return `msg_user_${crypto.randomUUID().replaceAll("-", "")}`;
-}
-
 function firstRow<T>(rows: T[]) {
   return rows[0] ?? null;
 }
-
-type TurnStartConvergenceGateway = NativeSessionGateway & {
-  acknowledgeTurnStartPersisted?: (input: { sessionId: string; turnId: string }) => Promise<void>;
-};
 
 function asRuntimeError(error: unknown) {
   return error instanceof RuntimeError ? error : null;
@@ -306,6 +296,8 @@ function mapThreadInterruptError(error: unknown, threadId: string): never {
 }
 
 export class ThreadService {
+  private readonly threadInputOrchestrator: ThreadInputOrchestrator;
+
   constructor(
     private readonly database: RuntimeDatabase,
     private readonly workspaceRegistry: WorkspaceRegistry,
@@ -313,15 +305,13 @@ export class ThreadService {
     private readonly sessionEventPublisher: SessionEventPublisher,
     private readonly nativeSessionGateway: NativeSessionGateway,
     private readonly now: () => Date = () => new Date(),
-  ) {}
-
-  private async acknowledgeTurnStartPersisted(sessionId: string, turnId: string) {
-    await (
-      this.nativeSessionGateway as TurnStartConvergenceGateway
-    ).acknowledgeTurnStartPersisted?.({
-      sessionId,
-      turnId,
-    });
+  ) {
+    this.threadInputOrchestrator = new ThreadInputOrchestrator(
+      this.database,
+      this.sessionEventPublisher,
+      this.nativeSessionGateway,
+      this.now,
+    );
   }
 
   async listThreads(workspaceId: string) {
@@ -403,7 +393,7 @@ export class ThreadService {
       deriveThreadTitle(input.content),
     );
     await this.startThreadSession(threadId);
-    await this.acceptThreadMessage(threadId, {
+    await this.threadInputOrchestrator.acceptThreadMessage(threadId, {
       client_message_id: input.client_request_id,
       content: input.content,
     });
@@ -430,7 +420,7 @@ export class ThreadService {
   ): Promise<{ thread: ThreadSummary; accepted_input: MessageProjection }> {
     let acceptedInput: MessageProjection;
     try {
-      acceptedInput = await this.acceptThreadMessage(threadId, {
+      acceptedInput = await this.threadInputOrchestrator.acceptThreadMessage(threadId, {
         client_message_id: input.client_request_id,
         content: input.content,
       });
@@ -936,191 +926,5 @@ export class ThreadService {
     })();
 
     return this.getThread(threadId);
-  }
-
-  private async acceptThreadMessage(
-    threadId: string,
-    input: {
-      client_message_id: string;
-      content: string;
-    },
-  ) {
-    const existingMessage = firstRow(
-      this.database.db
-        .select()
-        .from(messages)
-        .where(
-          and(
-            eq(messages.sessionId, threadId),
-            eq(messages.clientMessageId, input.client_message_id),
-          ),
-        )
-        .limit(1)
-        .all(),
-    );
-
-    if (existingMessage) {
-      if (existingMessage.content !== input.content) {
-        throw new RuntimeError(
-          409,
-          "message_idempotency_conflict",
-          "client message id conflicts with a different message payload",
-          {
-            session_id: threadId,
-            client_message_id: input.client_message_id,
-          },
-        );
-      }
-
-      return {
-        message_id: existingMessage.messageId,
-        session_id: existingMessage.sessionId,
-        role: existingMessage.role as MessageProjection["role"],
-        content: existingMessage.content,
-        created_at: existingMessage.createdAt,
-        source_item_type: existingMessage.sourceItemType as MessageProjection["source_item_type"],
-      } satisfies MessageProjection;
-    }
-
-    const session = firstRow(
-      this.database.db
-        .select()
-        .from(sessions)
-        .where(eq(sessions.sessionId, threadId))
-        .limit(1)
-        .all(),
-    );
-
-    if (!session) {
-      throw new RuntimeError(404, "session_not_found", "session was not found", {
-        session_id: threadId,
-      });
-    }
-
-    if (session.status !== "waiting_input") {
-      throw new RuntimeError(
-        409,
-        "session_invalid_state",
-        "session is not ready to accept messages",
-        {
-          session_id: threadId,
-          current_status: session.status,
-        },
-      );
-    }
-
-    const nativeTurn = await this.nativeSessionGateway.sendUserMessage({
-      sessionId: threadId,
-      content: input.content,
-    });
-    const userMessageId = generateMessageId();
-    const userCreatedAt = this.now().toISOString();
-
-    try {
-      this.database.sqlite.transaction(() => {
-        this.database.db
-          .insert(messages)
-          .values({
-            messageId: userMessageId,
-            sessionId: threadId,
-            role: "user",
-            content: input.content,
-            createdAt: userCreatedAt,
-            sourceItemType: "user_message",
-            clientMessageId: input.client_message_id,
-          })
-          .run();
-
-        this.sessionEventPublisher.appendSessionEvent({
-          sessionId: threadId,
-          eventType: "message.user",
-          occurredAt: userCreatedAt,
-          payload: {
-            message_id: userMessageId,
-            content: input.content,
-          },
-        });
-
-        this.database.db
-          .update(sessions)
-          .set({
-            status: "running",
-            updatedAt: userCreatedAt,
-            lastMessageAt: userCreatedAt,
-            currentTurnId: nativeTurn.turnId,
-            pendingAssistantMessageId: null,
-            appSessionOverlayState: "open",
-          })
-          .where(eq(sessions.sessionId, threadId))
-          .run();
-
-        this.database.db
-          .update(workspaces)
-          .set({
-            updatedAt: userCreatedAt,
-          })
-          .where(eq(workspaces.workspaceId, session.workspaceId))
-          .run();
-
-        this.sessionEventPublisher.appendSessionStatusChangedEvent({
-          sessionId: threadId,
-          occurredAt: userCreatedAt,
-          fromStatus: "waiting_input",
-          toStatus: "running",
-        });
-      })();
-      await this.acknowledgeTurnStartPersisted(threadId, nativeTurn.turnId);
-    } catch (error) {
-      try {
-        this.markThreadRecoveryPending(
-          threadId,
-          session.workspaceId,
-          userCreatedAt,
-          nativeTurn.turnId,
-        );
-      } catch {
-        // Keep the original post-native persistence failure as the primary error.
-      }
-      throw error;
-    }
-
-    return {
-      message_id: userMessageId,
-      session_id: threadId,
-      role: "user",
-      content: input.content,
-      created_at: userCreatedAt,
-      source_item_type: "user_message",
-    } satisfies MessageProjection;
-  }
-
-  private markThreadRecoveryPending(
-    threadId: string,
-    workspaceId: string,
-    updatedAt: string,
-    currentTurnId: string,
-  ) {
-    this.database.sqlite.transaction(() => {
-      this.database.db
-        .update(sessions)
-        .set({
-          status: "running",
-          updatedAt,
-          lastMessageAt: updatedAt,
-          currentTurnId,
-          pendingAssistantMessageId: null,
-          appSessionOverlayState: "recovery_pending",
-        })
-        .where(eq(sessions.sessionId, threadId))
-        .run();
-
-      this.database.db
-        .update(workspaces)
-        .set({
-          updatedAt,
-        })
-        .where(eq(workspaces.workspaceId, workspaceId))
-        .run();
-    })();
   }
 }
